+++ toys/StackExchange.Redis.Server/RedisServer.cs
@@ -15,6 +15,8 @@ namespace StackExchange.Redis.Server
 {
     public abstract partial class RedisServer : RespServer
     {
+        public const int DefaultDatabaseCount = 16;
+
         // non-trivial wildcards not implemented yet!
         public static bool IsMatch(string pattern, string key) =>
             pattern == "*" || string.Equals(pattern, key, StringComparison.OrdinalIgnoreCase);
@@ -123,7 +125,7 @@ namespace StackExchange.Redis.Server
             return endpoint;
         }
 
-        protected RedisServer(EndPoint endpoint = null, int databases = 16, TextWriter output = null) : base(output)
+        protected RedisServer(EndPoint endpoint = null, int databases = DefaultDatabaseCount, TextWriter output = null) : base(output)
         {
             DefaultEndPoint = endpoint ??= new IPEndPoint(IPAddress.Loopback, 6379);
             _nodes.TryAdd(endpoint, new Node(this, endpoint, NodeFlags.None));
@@ -1113,7 +1115,7 @@ namespace StackExchange.Redis.Server
         private static readonly Version s_DefaultServerVersion = new(1, 0, 0);
 
         private string _versionString;
-        private string VersionString => _versionString;
+        public string VersionString => _versionString;
         private static string FormatVersion(Version v)
         {
             var sb = new StringBuilder().Append(v.Major).Append('.').Append(v.Minor);
@@ -1153,7 +1155,6 @@ namespace StackExchange.Redis.Server
             switch (section)
             {
                 case "Server":
-                    var v = RedisVersion;
                     AddHeader().Append("redis_version:").AppendLine(VersionString)
                         .Append("redis_mode:").Append(ModeString).AppendLine()
                         .Append("os:").Append(Environment.OSVersion).AppendLine()
@@ -1259,8 +1260,20 @@ namespace StackExchange.Redis.Server
             var raw = request.GetValue(1);
             if (!raw.TryParse(out int db)) return TypedRedisValue.Error("ERR invalid DB index");
             if (db < 0 || db >= Databases) return TypedRedisValue.Error("ERR DB index is out of range");
+            if (db != 0 && !SupportMultiDb(out var err)) return TypedRedisValue.Error(err);
             client.Database = db;
             return TypedRedisValue.OK;
+        }
+
+        protected virtual bool SupportMultiDb(out string err)
+        {
+            if (ServerType is ServerType.Cluster)
+            {
+                err = "ERR SELECT is not allowed in cluster mode";
+                return false;
+            }
+            err = "";
+            return true;
         }
 
         private static readonly DateTime UnixEpoch = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);